requirements.txt
beautifulsoup4==4.9.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
faktory==0.4.0
Flask==1.1.2
Flask-Cors==3.0.8
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
python-dateutil==2.8.1
redis==3.5.2
requests==2.23.0
six==1.14.0
soupsieve==2.0.1
urllib3==1.25.9
urlparser==0.1.2
Werkzeug==1.0.1
